+++ CHANGELOG.md
@@ -1,15 +1,25 @@
-# 0.0.4 - 2018/01/10
+# Version 1.0.0 - 2024-07-02
+* Updates
+    * Replace auth mechanism by a Personal access token preset
+    * Python 3 support
+    * Add a state selector on the issues connector
+* New components
+    * Add a connector to fetch pull requests on specified repositories
+    * Add a connector to search pull issues based on a search query
+    * Add a connector to search pull requests based on predefined filters
+
+# Version 0.0.4 - 2018-01-10
 
 * Switch to using a code env
 
-# 0.0.3 - ?
+# Version 0.0.3 - ?
 
-# 0.0.2 - 2016/04/07
+# 0.0.2 - 2016-04-07
 
 * Add issue body
 * Add option to fetch comments
 * Fix "learn more" URL
 
-# 0.0.1 - 2015/09/28
+# Version 0.0.1 - 2015-09-28
 
 Initial release